michaels/requirements.txt
albumentations==0.5.2
anaconda==0.0.1.1
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.6.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==21.5b1
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
c3d==0.3.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
graphviz==0.16
h5py==3.2.1
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imgaug==0.4.0
ipykernel==5.5.4
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
notebook==6.3.0
numexpr==2.7.3
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
opensim===4.2-
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pathspec==0.8.1
pdfarranger==1.4.2
pexpect==4.6.0
pickleshare==0.7.5
pikepdf==1.10.3+dfsg
Pillow==7.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.8.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.9.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.16.0
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
ssh-import-id==5.10
systemd-python==234
tables==3.6.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
torch==1.9.0+cu111
torchaudio==0.9.0
torchgeometry==0.1.2
torchvision==0.10.0+cu111
torchviz==0.0.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vgg-pytorch==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
stable-baselines3==1.5.0
gym==0.24.0
tqdm==4.60.0

